Refactor controller-related symbol names
